Version 1.1.6

+++ app.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 const program = require('commander');
-const cmd = require('./Commands/index1')
+const cmd = require('./Commands/index')
 const pkg = require('./package.json')
 const {cencrypt,cdecrypt} = require('./CryptoAlgorithms/caesarChiper');
 const ceasarBrute = require('./CryptoAlgorithms/caesarBruteForce');
@@ -10,15 +10,13 @@ const {generateKeys,deleteKeys} = require('./ModernEncryption/keyGenerator');
 const {exportKey, importKey} = require('./ModernEncryption/keyPathChange');
 const {encryptMessage,decryptMessage} = require('./ModernEncryption/sendEncryptedMessage');
 
-program
-  .version(pkg.version)
-  .description('Simple Encryption Algorithms')
+cmd.setVersion(pkg.version)
 
 program
   .command('caesar <command> <key> <text>')
   .description('Encrypts or Decrypts given text with Caesar chiper\nOptions:\nencrypt - Encrypt Plaintext with Caesar method\ndecrypt - Decrypt Chipertext with Caesar method\nkey - number of letters to shift\n')
 
-if(process.argv.slice(5).length != 0){ //check if 5th argument exists to run the below code. References are on references.txt
+if(process.argv[2]=='caesar' && (process.argv[3]=='encrypt' || process.argv[3]=='decrypt')){ //check if 5th argument exists to run the below code. References are on references.txt
     cmd.parse(process.argv)
     cmd.description('Encrypts or Decrypts given text with Caesar chiper\nCommands:\nencrypt - Encrypt Plaintext with Caesar method\ndecrypt - Decrypt Chipertext with Caesar method\n')
     cmd.set('caesar <command> <key> [text]').action((command,key,text)=>{
@@ -31,21 +29,19 @@ if(process.argv.slice(5).length != 0){ //check if 5th argument exists to run the
         console.log('Unknown Command at: caesar ' + command);
         process.exit();
       }
-    });
+    }).description('test');
     cmd.close()
-}else{
-  program
-    .command('caesar <brute-force> <chipertext>')
-    .description('Brute forces decrypts a chipertext with Ceasar method\n')
-    .action((bf,chipertext)=>{
-      ceasarBrute(chipertext)
-    })
 }
-
-program
-  .command('tap-code <command> <plaintext>')
-  .description('Tap-Code Encryption\nOptions:\nencode - Encodes Plaintext with Tap-Code method\ndecode = Decodes Ciphertext with Tap-Code method\n')
-  .action((command,plaintext)=>{
+else if(process.argv[2]=='caesar' && process.argv[3]=='brute-force'){
+  cmd.parse(process.argv)
+  cmd.description('Brute forces decrypts a chipertext with Ceasar method\n')
+  cmd.set('caesar brute-force <chipertext>').action((chipertext)=>{
+    ceasarBrute(chipertext)
+  })
+}
+else if(process.argv[2]=='tap-code'){
+  cmd.parse(process.argv)
+  cmd.set('tap-code <command> <plaintext>').action((command,plaintext)=>{
     if(command == 'encode'){
       tencrypt(plaintext);
     }else if(command=='decode'){
@@ -54,11 +50,10 @@ program
       console.log('Unknown Command at: tap-code ' + command);
     }
   })
-
-program
-  .command('beale <command> <file> <plaintext>')
-  .description('Beale Encryption\nOptions:\nencrypt - Encrypts a plaintext with Beale Cipher Method\ndecrypt - Decrypts a ciphertext whith Beale Cipher Method\nfile - specifies a file that is the key to decrypt the ciphertext\n')
-  .action((command,file,plaintext)=>{
+}
+else if(process.argv[2]=='beale'){
+  cmd.parse(process.argv)
+  cmd.set('beale <command> <file> <plaintext>').action((command,file,plaintext)=> {
     if(command=='encrypt'){
       bencrypt(file,plaintext);
     }else if(command == 'decrypt'){
@@ -67,47 +62,41 @@ program
       console.log('Unknown Command at: beale ' + command);
     }
   })
-
-program
-  .command('create-user <name>')
-  .description('')
-  .action(name=>{
-    generateKeys(name)
+}
+else if(process.argv[2].includes('user')){
+  cmd.parse(process.argv)
+  cmd.set('<command> <name>').action((command,name)=> {
+    if(command=='create-user'){
+      generateKeys(name)
+    }
+    else if(command==='delete-user'){
+      deleteKeys(name)
+    }
   })
-
-program
-  .command('delete-user <name>')
-  .description('')
-  .action(name=>{
-    deleteKeys(name)
-  })
-
-program
-  .command('export-key <public|private> <name> [file]')
-  .description('')
-  .action((visibility,name,file)=>{
+}
+else if(process.argv[2]=='export-key'){
+  cmd.parse(process.argv)
+  cmd.set('export-key <public|private> <name> [file]').action((visibility,name,file)=> {
     exportKey(visibility,name,file)
   })
-
-program
-  .command('import-key <name> <path>')
-  .description('')
-  .action((name,path)=>{
-    importKey(name,path);
+}
+else if(process.argv[2]=='import-key'){
+  cmd.parse(process.argv)
+  cmd.set('import-key <name> <path>').action((name,path)=>{
+    importKey(name,path)
   })
-
-program
-  .command('write-message <name> <message> [file]')
-  .description('')
-  .action((name,message,file)=>{
+}
+else if(process.argv[2]=='write-message'){
+  cmd.parse(process.argv)
+  cmd.set('write-message <name> <message> [file]').action((name,message,file)=>{
     encryptMessage(name,message,file)
   })
-
-program
-  .command('read-message <encrypted-message>')
-  .description('')
-  .action(message=>{
+}
+else if(process.argv[2]=='read-message'){
+  cmd.parse(process.argv)
+  cmd.set('read-message <encrypted-message>').action((message)=>{
     decryptMessage(message)
   })
-
-program.parse(process.argv); //parsing the array of line arguments
+}else {
+  console.log('Unknown command at: ' + process.argv[2]);
+}
